Load feature PKGBUILD if any
For the original package, PGKBUILD is used by default.
If there is a feature, e.g, @foobar, we will check if
there is a file PKGBUILD@foobar, and load that file.
This way we do not need to edit the original file, bc.
everything can be overwritten in the feature file.

Is that good?