!bug for with an array of empty strings is buggy
When the input of a :for loop contains some empty string,
that string can make the program buggy. We have two ways
to avoid this issue

1. Remove quoting: for i in ${arr[@]-}; do something; done
2. Strict check: [[ -n "${arr[@]-}" ]] && for in in "${arr[@]}"

In #makepkg there are different quoting style. Though the
quoting is preferred, (assume that) we only have non-space
package names. Fixing by the 2nd way is a head-ache!